+++ java/org/apache/catalina/startup/HostConfig.java
@@ -663,9 +663,9 @@ public class HostConfig implements LifecycleListener {
                 if (!docBase.getCanonicalFile().toPath().startsWith(host.getAppBaseFile().toPath())) {
                     isExternal = true;
                     deployedApp.redeployResources.put(
-                            contextXml.getAbsolutePath(), Long.valueOf(contextXml.lastModified()));
+                            contextXml.getAbsolutePath(), contextXml.lastModified());
                     deployedApp.redeployResources.put(
-                            docBase.getAbsolutePath(), Long.valueOf(docBase.lastModified()));
+                            docBase.getAbsolutePath(), docBase.lastModified());
                     if (docBase.getAbsolutePath().toLowerCase(Locale.ENGLISH).endsWith(".war")) {
                         isExternalWar = true;
                     }
@@ -715,7 +715,7 @@ public class HostConfig implements LifecycleListener {
             if (isExternalWar) {
                 if (unpackWAR) {
                     deployedApp.redeployResources.put(
-                            expandedDocBase.getAbsolutePath(), Long.valueOf(expandedDocBase.lastModified()));
+                            expandedDocBase.getAbsolutePath(), expandedDocBase.lastModified());
                     addWatchedResources(deployedApp, expandedDocBase.getAbsolutePath(), context);
                 } else {
                     addWatchedResources(deployedApp, null, context);
@@ -726,15 +726,15 @@ public class HostConfig implements LifecycleListener {
                     File warDocBase = new File(expandedDocBase.getAbsolutePath() + ".war");
                     if (warDocBase.exists()) {
                         deployedApp.redeployResources.put(
-                                warDocBase.getAbsolutePath(), Long.valueOf(warDocBase.lastModified()));
+                                warDocBase.getAbsolutePath(), warDocBase.lastModified());
                     } else {
                         // Trigger a redeploy if a WAR is added
-                        deployedApp.redeployResources.put(warDocBase.getAbsolutePath(), Long.valueOf(0));
+                        deployedApp.redeployResources.put(warDocBase.getAbsolutePath(), 0L);
                     }
                 }
                 if (unpackWAR) {
                     deployedApp.redeployResources.put(
-                            expandedDocBase.getAbsolutePath(), Long.valueOf(expandedDocBase.lastModified()));
+                            expandedDocBase.getAbsolutePath(), expandedDocBase.lastModified());
                     addWatchedResources(deployedApp, expandedDocBase.getAbsolutePath(), context);
                 } else {
                     addWatchedResources(deployedApp, null, context);
@@ -743,7 +743,7 @@ public class HostConfig implements LifecycleListener {
                     // For external docBases, the context.xml will have been
                     // added above.
                     deployedApp.redeployResources.put(
-                            contextXml.getAbsolutePath(), Long.valueOf(contextXml.lastModified()));
+                            contextXml.getAbsolutePath(), contextXml.lastModified());
                 }
             }
             // Add the global redeploy resources (which are never deleted) at
@@ -757,7 +757,7 @@ public class HostConfig implements LifecycleListener {
 
         if (log.isInfoEnabled()) {
             log.info(sm.getString("hostConfig.deployDescriptor.finished",
-                    contextXml.getAbsolutePath(), Long.valueOf(System.currentTimeMillis() - startTime)));
+                    contextXml.getAbsolutePath(), System.currentTimeMillis() - startTime));
         }
     }
 
@@ -1002,15 +1002,15 @@ public class HostConfig implements LifecycleListener {
 
         try {
             // Populate redeploy resources with the WAR file
-            deployedApp.redeployResources.put(war.getAbsolutePath(), Long.valueOf(war.lastModified()));
+            deployedApp.redeployResources.put(war.getAbsolutePath(), war.lastModified());
 
             if (deployThisXML && xml.exists() && copyThisXml) {
-                deployedApp.redeployResources.put(xml.getAbsolutePath(), Long.valueOf(xml.lastModified()));
+                deployedApp.redeployResources.put(xml.getAbsolutePath(), xml.lastModified());
             } else {
                 // In case an XML file is added to the config base later
                 deployedApp.redeployResources.put(
                         (new File(host.getConfigBaseFile(), cn.getBaseName() + ".xml")).getAbsolutePath(),
-                        Long.valueOf(0));
+                    0L);
             }
 
             Class<?> clazz = Class.forName(host.getConfigClass());
@@ -1034,10 +1034,10 @@ public class HostConfig implements LifecycleListener {
             }
             if (unpackWAR && context.getDocBase() != null) {
                 File docBase = new File(host.getAppBaseFile(), cn.getBaseName());
-                deployedApp.redeployResources.put(docBase.getAbsolutePath(), Long.valueOf(docBase.lastModified()));
+                deployedApp.redeployResources.put(docBase.getAbsolutePath(), docBase.lastModified());
                 addWatchedResources(deployedApp, docBase.getAbsolutePath(), context);
                 if (deployThisXML && !copyThisXml && (xmlInWar || xml.exists())) {
-                    deployedApp.redeployResources.put(xml.getAbsolutePath(), Long.valueOf(xml.lastModified()));
+                    deployedApp.redeployResources.put(xml.getAbsolutePath(), xml.lastModified());
                 }
             } else {
                 // Passing null for docBase means that no resources will be
@@ -1053,7 +1053,7 @@ public class HostConfig implements LifecycleListener {
 
         if (log.isInfoEnabled()) {
             log.info(sm.getString("hostConfig.deployWar.finished",
-                    war.getAbsolutePath(), Long.valueOf(System.currentTimeMillis() - startTime)));
+                    war.getAbsolutePath(), System.currentTimeMillis() - startTime));
         }
     }
 
@@ -1191,23 +1191,23 @@ public class HostConfig implements LifecycleListener {
 
             // Fake re-deploy resource to detect if a WAR is added at a later
             // point
-            deployedApp.redeployResources.put(dir.getAbsolutePath() + ".war", Long.valueOf(0));
-            deployedApp.redeployResources.put(dir.getAbsolutePath(), Long.valueOf(dir.lastModified()));
+            deployedApp.redeployResources.put(dir.getAbsolutePath() + ".war", 0L);
+            deployedApp.redeployResources.put(dir.getAbsolutePath(), dir.lastModified());
             if (deployThisXML && xml.exists()) {
                 if (copyThisXml) {
-                    deployedApp.redeployResources.put(xmlCopy.getAbsolutePath(), Long.valueOf(xmlCopy.lastModified()));
+                    deployedApp.redeployResources.put(xmlCopy.getAbsolutePath(), xmlCopy.lastModified());
                 } else {
-                    deployedApp.redeployResources.put(xml.getAbsolutePath(), Long.valueOf(xml.lastModified()));
+                    deployedApp.redeployResources.put(xml.getAbsolutePath(), xml.lastModified());
                     // Fake re-deploy resource to detect if a context.xml file is
                     // added at a later point
-                    deployedApp.redeployResources.put(xmlCopy.getAbsolutePath(), Long.valueOf(0));
+                    deployedApp.redeployResources.put(xmlCopy.getAbsolutePath(), 0L);
                 }
             } else {
                 // Fake re-deploy resource to detect if a context.xml file is
                 // added at a later point
-                deployedApp.redeployResources.put(xmlCopy.getAbsolutePath(), Long.valueOf(0));
+                deployedApp.redeployResources.put(xmlCopy.getAbsolutePath(), 0L);
                 if (!xml.exists()) {
-                    deployedApp.redeployResources.put(xml.getAbsolutePath(), Long.valueOf(0));
+                    deployedApp.redeployResources.put(xml.getAbsolutePath(), 0L);
                 }
             }
             addWatchedResources(deployedApp, dir.getAbsolutePath(), context);
@@ -1220,7 +1220,7 @@ public class HostConfig implements LifecycleListener {
 
         if( log.isInfoEnabled() ) {
             log.info(sm.getString("hostConfig.deployDir.finished",
-                    dir.getAbsolutePath(), Long.valueOf(System.currentTimeMillis() - startTime)));
+                    dir.getAbsolutePath(), System.currentTimeMillis() - startTime));
         }
     }
 
@@ -1272,7 +1272,7 @@ public class HostConfig implements LifecycleListener {
                         resource.getAbsolutePath() + "'");
             }
             app.reloadResources.put(resource.getAbsolutePath(),
-                    Long.valueOf(resource.lastModified()));
+                resource.lastModified());
         }
     }
 
@@ -1283,7 +1283,7 @@ public class HostConfig implements LifecycleListener {
                 new File(getConfigBaseName(), Constants.HostContextXml);
         if (hostContextXml.isFile()) {
             app.redeployResources.put(hostContextXml.getAbsolutePath(),
-                    Long.valueOf(hostContextXml.lastModified()));
+                hostContextXml.lastModified());
         }
 
         // Redeploy resources in CATALINA_BASE/conf are never deleted
@@ -1291,7 +1291,7 @@ public class HostConfig implements LifecycleListener {
                 returnCanonicalPath(Constants.DefaultContextXml);
         if (globalContextXml.isFile()) {
             app.redeployResources.put(globalContextXml.getAbsolutePath(),
-                    Long.valueOf(globalContextXml.lastModified()));
+                globalContextXml.lastModified());
         }
     }
 
@@ -1319,8 +1319,7 @@ public class HostConfig implements LifecycleListener {
                 log.debug("Checking context[" + app.name +
                         "] redeploy resource " + resource);
             }
-            long lastModified =
-                    app.redeployResources.get(resources[i]).longValue();
+            long lastModified = app.redeployResources.get(resources[i]);
             if (resource.exists() || lastModified == 0) {
                 // File.lastModified() has a resolution of 1s (1000ms). The last
                 // modified time has to be more than 1000ms ago to ensure that
@@ -1332,7 +1331,7 @@ public class HostConfig implements LifecycleListener {
                     if (resource.isDirectory()) {
                         // No action required for modified directory
                         app.redeployResources.put(resources[i],
-                                Long.valueOf(resource.lastModified()));
+                            resource.lastModified());
                     } else if (app.hasDescriptor &&
                             resource.getName().toLowerCase(
                                     Locale.ENGLISH).endsWith(".war")) {
@@ -1355,7 +1354,7 @@ public class HostConfig implements LifecycleListener {
                         }
                         // Update times
                         app.redeployResources.put(resources[i],
-                                Long.valueOf(resource.lastModified()));
+                            resource.lastModified());
                         app.timestamp = System.currentTimeMillis();
                         boolean unpackWAR = unpackWARs;
                         if (unpackWAR && context instanceof StandardContext) {
@@ -1400,7 +1399,7 @@ public class HostConfig implements LifecycleListener {
             if (log.isDebugEnabled()) {
                 log.debug("Checking context[" + app.name + "] reload resource " + resource);
             }
-            long lastModified = app.reloadResources.get(s).longValue();
+            long lastModified = app.reloadResources.get(s);
             // File.lastModified() has a resolution of 1s (1000ms). The last
             // modified time has to be more than 1000ms ago to ensure that
             // modifications that take place in the same second are not
@@ -1418,7 +1417,7 @@ public class HostConfig implements LifecycleListener {
                 // Update times. More than one file may have been updated. We
                 // don't want to trigger a series of reloads.
                 app.reloadResources.put(s,
-                        Long.valueOf(resource.lastModified()));
+                    resource.lastModified());
             }
             app.timestamp = System.currentTimeMillis();
         }
@@ -1775,7 +1774,7 @@ public class HostConfig implements LifecycleListener {
                 docBase = new File(host.getAppBaseFile(), context.getDocBase());
             }
             deployedApp.redeployResources.put(docBase.getAbsolutePath(),
-                    Long.valueOf(docBase.lastModified()));
+                docBase.lastModified());
             if (docBase.getAbsolutePath().toLowerCase(Locale.ENGLISH).endsWith(".war")) {
                 isWar = true;
             }
@@ -1790,7 +1789,7 @@ public class HostConfig implements LifecycleListener {
         if (isWar && unpackWAR) {
             File docBase = new File(host.getAppBaseFile(), context.getBaseName());
             deployedApp.redeployResources.put(docBase.getAbsolutePath(),
-                        Long.valueOf(docBase.lastModified()));
+                docBase.lastModified());
             addWatchedResources(deployedApp, docBase.getAbsolutePath(), context);
         } else {
             addWatchedResources(deployedApp, null, context);
