+++ java/org/apache/coyote/http2/ConnectionSettingsBase.java
@@ -52,12 +52,12 @@ abstract class ConnectionSettingsBase<T extends Throwable> {
     ConnectionSettingsBase(String connectionId) {
         this.connectionId = connectionId;
         // Set up the defaults
-        current.put(Setting.HEADER_TABLE_SIZE,      Long.valueOf(DEFAULT_HEADER_TABLE_SIZE));
-        current.put(Setting.ENABLE_PUSH,            Long.valueOf(DEFAULT_ENABLE_PUSH ? 1 : 0));
-        current.put(Setting.MAX_CONCURRENT_STREAMS, Long.valueOf(DEFAULT_MAX_CONCURRENT_STREAMS));
-        current.put(Setting.INITIAL_WINDOW_SIZE,    Long.valueOf(DEFAULT_INITIAL_WINDOW_SIZE));
-        current.put(Setting.MAX_FRAME_SIZE,         Long.valueOf(DEFAULT_MAX_FRAME_SIZE));
-        current.put(Setting.MAX_HEADER_LIST_SIZE,   Long.valueOf(DEFAULT_MAX_HEADER_LIST_SIZE));
+        current.put(Setting.HEADER_TABLE_SIZE, (long) DEFAULT_HEADER_TABLE_SIZE);
+        current.put(Setting.ENABLE_PUSH, (long) (DEFAULT_ENABLE_PUSH ? 1 : 0));
+        current.put(Setting.MAX_CONCURRENT_STREAMS, DEFAULT_MAX_CONCURRENT_STREAMS);
+        current.put(Setting.INITIAL_WINDOW_SIZE, (long) DEFAULT_INITIAL_WINDOW_SIZE);
+        current.put(Setting.MAX_FRAME_SIZE, (long) DEFAULT_MAX_FRAME_SIZE);
+        current.put(Setting.MAX_HEADER_LIST_SIZE, DEFAULT_MAX_HEADER_LIST_SIZE);
     }
 
 
@@ -135,11 +135,11 @@ abstract class ConnectionSettingsBase<T extends Throwable> {
 
     private synchronized long getMin(Setting setting) {
         Long pendingValue = pending.get(setting);
-        long currentValue = current.get(setting).longValue();
+        long currentValue = current.get(setting);
         if (pendingValue == null) {
             return currentValue;
         } else {
-            return Long.min(pendingValue.longValue(), currentValue);
+            return Long.min(pendingValue, currentValue);
         }
     }
 
@@ -156,11 +156,11 @@ abstract class ConnectionSettingsBase<T extends Throwable> {
 
     private synchronized long getMax(Setting setting) {
         Long pendingValue = pending.get(setting);
-        long currentValue = current.get(setting).longValue();
+        long currentValue = current.get(setting);
         if (pendingValue == null) {
             return currentValue;
         } else {
-            return Long.max(pendingValue.longValue(), currentValue);
+            return Long.max(pendingValue, currentValue);
         }
     }
